Removed handlers and validators and put them in a separate package
