+++ .env.example
@@ -1,2 +1,8 @@
 BSKY_HANDLE='HANDLE'
 BSKY_PASSWORD='APP_PASSWORD'
+DB_PORT=3306
+DB_DATABASE=db
+DB_USERNAME=user
+DB_PASSWORD=password
+DB_HOST=localhost
+DEBUG_LOG_ACTIVE=true
