+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@atproto/api": "^0.6.23",
     "atproto-firehose": "^0.2.2",
+    "bsky-event-handlers": "^0.0.2",
     "date-fns": "^2.30.0",
     "mysql": "^2.18.1",
     "mysql2": "^3.6.5",
