--- src/utils/post-details-utils.ts
@@ -1,5 +1,0 @@
-import {PostDetails} from "./types.ts";
-
-export function getPosterDID(postDetails: PostDetails){
-    return (postDetails.uri.match(/did:[^\/]*/) || [])[0];
-}
